// react-techwatch/src/models/Cart.ts
export interface cartinterface {
    // uuid: uuid.v4(),
    // id: payload.id,
    // name: payload.name,
    // price: parseFloat(payload.price),
    // totalprice: parseFloat(payload.price) * 1,
    // img: payload.img,
    // qty: 1
    uuid: string;
    id: number;
    name: string;
    price: number;
    totalprice: number;
    img: string
    qty: number
}